refactor(content): extract selection mode and status-url helpers
Selection mode (~180 lines: overlay marks, floating bar, enter/exit) moves
out of injector.ts into its own content module, and the shared permalink
helpers (normalizeStatusUrl/getStatusUrl) into a status-url leaf both
import. Clean DAG, no cycle: injector -> selection -> status-url, injector
-> status-url. esbuild inlines both back into the injector IIFE (bundle
size unchanged). Pure move — typecheck, build, and 194 tests green.
